+++ example/lib/ui/example/useStage/basic.dart
@@ -12,6 +12,15 @@ import 'package:flutter/material.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 import '../../Item.dart';
 
+/*
+   the most common usage,
+   child wrap ListView,GridView,CustomView,Widget,
+   RefreshConfiguration is a global setting,just  like theme,All refreshers under the RefreshConfiguration subtree will refer to its properties,
+   in this example,I gave RefreshConfiguration a property headerBuilder,The default header indicator for the four refreshers is it.
+   If you use almost the same page indicator, using Refresh Configuration can greatly reduce the complexity of your work.
+
+*/
+
 class BasicExample extends StatefulWidget {
   @override
   State<StatefulWidget> createState() {
@@ -25,21 +34,30 @@ class _BasicExampleState extends State<BasicExample>
   RefreshController _refreshController1 =
       RefreshController(initialRefresh: true);
   RefreshController _refreshController2 =
-  RefreshController(initialRefresh: true);
+      RefreshController(initialRefresh: true);
   RefreshController _refreshController3 =
+      RefreshController(initialRefresh: true);
+  RefreshController _refreshController4 =
   RefreshController(initialRefresh: true);
+
 //  int pageIndex = 0;
-  List<String> data1 = [],data2 = [],data3 = [];
+  List<String> data1 = [], data2 = [], data3 = [];
   TabController _tabController;
 
   @override
   void initState() {
     // TODO: implement initState
-    _tabController = TabController(length: 3, vsync: this);
-    _tabController.addListener((){
-
-    });
-
+    _tabController = TabController(length: 4, vsync: this);
+    _tabController.addListener(() {});
+    for (int i = 0; i < 10; i++) {
+      data1.add("Item $i");
+    }
+    for (int i = 0; i < 10; i++) {
+      data2.add("Item $i");
+    }
+    for (int i = 0; i < 10; i++) {
+      data3.add("Item $i");
+    }
     super.initState();
   }
 
@@ -49,8 +67,7 @@ class _BasicExampleState extends State<BasicExample>
     super.didChangeDependencies();
   }
 
-
-  void _onRefresh(RefreshController controller,List<String> data) async {
+  void _onRefresh(RefreshController controller, List<String> data) async {
     //monitor fetch data from network
     await Future.delayed(Duration(milliseconds: 1000));
 
@@ -70,7 +87,7 @@ class _BasicExampleState extends State<BasicExample>
       */
   }
 
-  void _onLoading(RefreshController controller,List<String> data) async {
+  void _onLoading(RefreshController controller, List<String> data) async {
     //monitor fetch data from network
     await Future.delayed(Duration(milliseconds: 2000));
     for (int i = 0; i < 10; i++) {
@@ -142,6 +159,9 @@ class _BasicExampleState extends State<BasicExample>
               ),
               Tab(
                 text: "CustomScrollView",
+              ),
+              Tab(
+                text: "其他组件",
               )
             ],
           ),
@@ -153,10 +173,10 @@ class _BasicExampleState extends State<BasicExample>
               child: buildList(),
               controller: _refreshController1,
               enablePullUp: true,
-              onRefresh: (){
-                _onRefresh(_refreshController1,data1);
+              onRefresh: () {
+                _onRefresh(_refreshController1, data1);
               },
-              onLoading: (){
+              onLoading: () {
                 _onLoading(_refreshController1, data1);
               },
             ),
@@ -164,10 +184,10 @@ class _BasicExampleState extends State<BasicExample>
               child: buildGrid(),
               controller: _refreshController2,
               enablePullUp: true,
-              onRefresh: (){
-                _onRefresh(_refreshController2,data2);
+              onRefresh: () {
+                _onRefresh(_refreshController2, data2);
               },
-              onLoading: (){
+              onLoading: () {
                 _onLoading(_refreshController2, data2);
               },
             ),
@@ -175,19 +195,53 @@ class _BasicExampleState extends State<BasicExample>
               child: buildCustom(),
               enablePullUp: true,
               controller: _refreshController3,
-              onRefresh: (){
-                _onRefresh(_refreshController3,data3);
+              onRefresh: () {
+                _onRefresh(_refreshController3, data3);
               },
-              onLoading: (){
+              onLoading: () {
                 _onLoading(_refreshController3, data3);
               },
+            ),
+            SmartRefresher(
+              child: Container(
+                height: 1000.0,
+                child: Column(
+                  children: <Widget>[
+                    Container(
+                      color: Colors.redAccent,
+                      height: 200.0,
+                    ),
+                    Text("标题"),
+                    Container(
+                      color: Colors.redAccent,
+                      height: 200.0,
+                    ),
+                    Text("标题"),
+                    Container(
+                      color: Colors.redAccent,
+                      height: 200.0,
+                    ),
+                    Text("标题"),
+                    Container(
+                      color: Colors.redAccent,
+                      height: 200.0,
+                    ),
+                    Text("标题"),
+                  ],
+                ),
+              ),
+              enablePullUp: true,
+              controller: _refreshController4,
+              onRefresh: () {},
+              onLoading: () {},
             )
           ],
         ),
       ),
-      headerBuilder: () => WaterDropMaterialHeader(backgroundColor: Theme.of(context).primaryColor,),
+      headerBuilder: () => WaterDropMaterialHeader(
+        backgroundColor: Theme.of(context).primaryColor,
+      ),
       footerTriggerDistance: 80.0,
-
     );
   }
 }