+++ example/lib/ui/test/Test4.dart
@@ -1,211 +1,230 @@
-// Copyright 2017 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
+import 'dart:async';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
-import '../../other/refresh_animatedlist.dart';
 
-class AnimatedListSample extends StatefulWidget {
+class Test4 extends StatefulWidget {
+  Test4({Key key}) : super(key: key);
+
   @override
-  _AnimatedListSampleState createState() => new _AnimatedListSampleState();
+  Test4State createState() => Test4State();
+
 }
 
-class _AnimatedListSampleState extends State<AnimatedListSample> {
-  final GlobalKey<AnimatedListState> _listKey =
-      new GlobalKey<AnimatedListState>();
-  ListModel<int> _list;
-  int _selectedItem;
-  int _nextItem; // The next item inserted when the user presses the '+' button.
+class Test4State extends State<Test4>
+    with AutomaticKeepAliveClientMixin, TickerProviderStateMixin {
+//  RefreshMode  refreshing = RefreshMode.idle;
+//  LoadMode loading = LoadMode.idle;
+  ValueNotifier<double> topOffsetLis = ValueNotifier(0.0);
+  ValueNotifier<double> bottomOffsetLis = ValueNotifier(0.0);
+  RefreshController _refreshController;
+  ScrollController _scrollController;
 
-  @override
-  void initState() {
-    super.initState();
-    _list = new ListModel<int>(
-      listKey: _listKey,
-      initialItems: <int>[0, 1, 2],
-      removedItemBuilder: _buildRemovedItem,
-    );
-    _nextItem = 3;
-  }
+  List<Widget> data = [];
 
-  // Used to build list items that haven't been removed.
-  Widget _buildItem(
-      BuildContext context, int index, Animation<double> animation) {
-    return new CardItem(
-      animation: animation,
-      item: _list[index],
-      selected: _selectedItem == _list[index],
-      onTap: () {
-        setState(() {
-          _selectedItem = _selectedItem == _list[index] ? null : _list[index];
-        });
-      },
-    );
-  }
+  //test #68
+  bool _enablePullUp = true, _enablePullDown = true;
 
-  // Used to build an item after it has been removed from the list. This method is
-  // needed because a removed item remains  visible until its animation has
-  // completed (even though it's gone as far this ListModel is concerned).
-  // The widget will be used by the [AnimatedListState.removeItem] method's
-  // [AnimatedListRemovedItemBuilder] parameter.
-  Widget _buildRemovedItem(
-      int item, BuildContext context, Animation<double> animation) {
-    return new CardItem(
-      animation: animation,
-      item: item,
-      selected: false,
-      // No gesture detector here: we don't want removed items to be interactive.
-    );
-  }
-
-  // Insert the "next item" into the list model.
-  void _insert() {
-    final int index =
-        _selectedItem == null ? _list.length : _list.indexOf(_selectedItem);
-    _list.insert(index, _nextItem++);
-  }
-
-  // Remove the selected item from the list model.
-  void _remove() {
-    if (_selectedItem != null) {
-      _list.removeAt(_list.indexOf(_selectedItem));
-      setState(() {
-        _selectedItem = null;
-      });
-    }
-  }
-
-  RefreshController _refreshController = RefreshController();
-
-  @override
-  Widget build(BuildContext context) {
-    return new MaterialApp(
-      home: new Scaffold(
-        appBar: new AppBar(
-          title: const Text('AnimatedList'),
-          actions: <Widget>[
-            new IconButton(
-              icon: const Icon(Icons.add_circle),
-              onPressed: _insert,
-              tooltip: 'insert a new item',
-            ),
-            new IconButton(
-              icon: const Icon(Icons.remove_circle),
-              onPressed: _remove,
-              tooltip: 'remove the selected item',
-            ),
-          ],
-        ),
-        body:  RefreshAnimatedList(
-          key: _listKey,
-          initialItemCount: _list.length,
-          itemBuilder: _buildItem,
-          enablePullUp: true,
-          refreshController: _refreshController,
-        ),
-      ),
-    );
-  }
-}
-
-/// Keeps a Dart List in sync with an AnimatedList.
-///
-/// The [insert] and [removeAt] methods apply to both the internal list and the
-/// animated list that belongs to [listKey].
-///
-/// This class only exposes as much of the Dart List API as is needed by the
-/// sample app. More list methods are easily added, however methods that mutate the
-/// list must make the same changes to the animated list in terms of
-/// [AnimatedListState.insertItem] and [AnimatedList.removeItem].
-class ListModel<E> {
-  ListModel({
-    @required this.listKey,
-    @required this.removedItemBuilder,
-    Iterable<E> initialItems,
-  })  : assert(listKey != null),
-        assert(removedItemBuilder != null),
-        _items = new List<E>.from(initialItems ?? <E>[]);
-
-  final GlobalKey<AnimatedListState> listKey;
-  final dynamic removedItemBuilder;
-  final List<E> _items;
-
-  AnimatedListState get _animatedList => listKey.currentState;
-
-  void insert(int index, E item) {
-    _items.insert(index, item);
-    _animatedList.insertItem(index);
-  }
-
-  E removeAt(int index) {
-    final E removedItem = _items.removeAt(index);
-    if (removedItem != null) {
-      _animatedList.removeItem(index,
-          (BuildContext context, Animation<double> animation) {
-        return removedItemBuilder(removedItem, context, animation);
-      });
-    }
-    return removedItem;
-  }
-
-  int get length => _items.length;
-
-  E operator [](int index) => _items[index];
-
-  int indexOf(E item) => _items.indexOf(item);
-}
-
-/// Displays its integer item as 'item N' on a Card whose color is based on
-/// the item's value. The text is displayed in bright green if selected is true.
-/// This widget's height is based on the animation parameter, it varies
-/// from 0 to 128 as the animation varies from 0.0 to 1.0.
-class CardItem extends StatelessWidget {
-  const CardItem(
-      {Key key,
-      @required this.animation,
-      this.onTap,
-      @required this.item,
-      this.selected: false})
-      : assert(animation != null),
-        assert(item != null && item >= 0),
-        assert(selected != null),
-        super(key: key);
-
-  final Animation<double> animation;
-  final VoidCallback onTap;
-  final int item;
-  final bool selected;
-
-  @override
-  Widget build(BuildContext context) {
-    TextStyle textStyle = Theme.of(context).textTheme.display1;
-    if (selected)
-      textStyle = textStyle.copyWith(color: Colors.lightGreenAccent[400]);
-    return new Padding(
-      padding: const EdgeInsets.all(2.0),
-      child: new SizeTransition(
-        axis: Axis.vertical,
-        sizeFactor: animation,
-        child: new GestureDetector(
-          behavior: HitTestBehavior.opaque,
-          onTap: onTap,
-          child: new SizedBox(
-            height: 128.0,
-            child: new Card(
-              color: Colors.primaries[item % Colors.primaries.length],
-              child: new Center(
-                child: new Text('Item $item', style: textStyle),
-              ),
+  void _getDatas() {
+    data.add(Row(
+      children: <Widget>[
+        FlatButton(
+            onPressed: () {
+              _refreshController.requestRefresh();
+            },
+            child: Text("请求刷新")),
+        FlatButton(
+            onPressed: () {
+              _refreshController.requestLoading();
+            },
+            child: Text("请求加载数据"))
+      ],
+    ));
+    for (int i = 0; i < 22; i++) {
+      data.add(GestureDetector(
+        child: Container(
+          color: Color.fromARGB(255, 250, 250, 250),
+          child: Card(
+            margin:
+            EdgeInsets.only(left: 10.0, right: 10.0, top: 5.0, bottom: 5.0),
+            child: Center(
+              child: Text('Data $i'),
             ),
           ),
         ),
-      ),
+        onTap: () {
+          _refreshController.requestRefresh();
+        },
+      ));
+    }
+  }
+
+  void enterRefresh() {
+    _refreshController.requestLoading();
+  }
+
+  void _onOffsetCallback(bool isUp, double offset) {
+    // if you want change some widgets state ,you should rewrite the callback
+    if (mounted) setState(() {});
+    if (isUp) {
+      print(offset);
+      bottomOffsetLis.value = offset;
+    } else {
+      print(offset);
+      topOffsetLis.value = offset;
+    }
+  }
+
+  @override
+  void initState() {
+    // TODO: implement initState
+
+    // for test #68 true-> false ->true
+//    Future.delayed(Duration(milliseconds: 3000), () {
+//      _enablePullDown = false;
+//      _enablePullUp = false;
+//      if (mounted) setState(() {});
+//    });
+//    Future.delayed(Duration(milliseconds: 6000), () {
+//      _enablePullDown = true;
+//      _enablePullUp = true;
+//      if (mounted) setState(() {});
+//    });
+
+//    // for test #68 false-> true ->false
+//    Future.delayed(Duration(milliseconds: 3000),(){
+//      _enablePullDown = false;
+//      _enablePullUp = true;
+//    if(mounted)
+//      setState(() {
+//
+//      });
+//    });
+//    Future.delayed(Duration(milliseconds: 6000),(){
+//      _enablePullDown = true;
+//      _enablePullUp = false;
+//    if(mounted)
+//      setState(() {
+//
+//      });
+//    });
+//    Future.delayed(Duration(milliseconds: 3000),(){
+//      _enablePullDown = true;
+//      _enablePullUp = false;
+//    if(mounted)
+//      setState(() {
+//
+//      });
+//    });
+//    Future.delayed(Duration(milliseconds: 6000),(){
+//      _enablePullDown = false;
+//      _enablePullUp = true;
+//    if(mounted)
+//      setState(() {
+//
+//      });
+//    });
+    _getDatas();
+    _refreshController = RefreshController(initialRefresh: true,initialLoadStatus: LoadStatus.noMore,initialRefreshStatus: RefreshStatus.refreshing);
+
+    super.initState();
+  }
+
+  Widget _headerCreate(BuildContext context, RefreshStatus mode) {
+    return Image.asset(
+      "images/animate.gif",
+      fit: BoxFit.fitWidth,
+      alignment: Alignment.topCenter,
     );
   }
+
+  @override
+  void dispose() {
+    // TODO: implement dispose
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: <Widget>[
+        Expanded(
+          child: SmartRefresher(
+            enablePullUp: _enablePullDown,
+            enablePullDown: _enablePullUp,
+            controller: _refreshController,
+            header: ClassicHeader(
+              iconPos: IconPosition.top,
+              height: 60.0,
+            ),
+            footer: ClassicFooter(
+              iconPos: IconPosition.top,
+              height: 60.0,
+            ),
+            onRefresh: ()  async{
+              print("onRefresh");
+              await Future.delayed(const Duration(milliseconds: 1000));
+              data.add(Container(
+                child: Card(),
+                height: 100.0,
+              ));
+              if (mounted) setState(() {});
+              _refreshController.refreshToIdle();
+//        Future.delayed(const Duration(milliseconds: 2009)).then((val) {
+////          data.add(Card());
+//
+//        });
+            },
+            child: ListView.builder(
+              itemExtent: 100.0,
+              physics: BouncingScrollPhysics(),
+              itemBuilder: (c,i) => data[i],
+              itemCount: data.length,
+            ),
+            onLoading: () {
+              print("onload");
+              Future.delayed(const Duration(milliseconds: 2000)).then((val) {
+                data.add(Card());
+                if (mounted) setState(() {});
+                _refreshController.loadComplete();
+              });
+            },
+          ),
+        )
+      ],
+    );
+  }
+
+  @override
+  // TODO: implement wantKeepAlive
+  bool get wantKeepAlive => false;
 }
 
-void main() {
-  runApp(new AnimatedListSample());
+class CirclePainter extends CustomClipper<Path> {
+  final double offset;
+  final bool up;
+
+  CirclePainter({this.offset, this.up});
+
+  @override
+  Path getClip(Size size) {
+    // TODO: implement getClip
+    final path = Path();
+    if (!up) path.moveTo(0.0, size.height);
+    path.cubicTo(
+        0.0,
+        up ? 0.0 : size.height,
+        size.width / 2,
+        up ? offset * 2.3 : size.height - offset * 2.3,
+        size.width,
+        up ? 0.0 : size.height);
+    path.close();
+    return path;
+  }
+
+  @override
+  bool shouldReclip(CustomClipper oldClipper) {
+    // TODO: implement shouldReclip
+    return oldClipper != this;
+  }
 }