+++ CHANGELOG.md
@@ -194,3 +194,33 @@ Notice: This version of the code changes much, Api too
 * Fix 1.4.5  default physics Bug  in Android simulation
 * Fix Problem: when enablePullDown or enablePullUp = false,it still can overScroll or underScroll when use ClampingScrollPhysics
 * Add maxOverScrollExtent and maxUnderScrollExtent in RefreshConfiguration
+
+## 1.4.7
+new Feature:
+* Add twoLevel refresh feature
+* Add linkHeader to link other place header
+
+SmartRefresher:
+* Remove headerInsertIndex
+* Fix ignore padding attr when child is BoxScrollView
+* add enableTwoLevel,onTwoLevel attr
+
+RefreshConfiguration:
+1. add enableScrollWhenTwoLevel,closeTwoLevelDistance for twoLevel setting
+
+RefreshController:
+* Add refreshToidle, twoLevelComplete new api
+* Add initalRefreshStatus,initalLoadStatus new parameter setting default value
+
+ClassicalIndicator:
+1. remove decoration
+2. add outerBuilder replace decoration
+3. add other attr for twoLevel
+
+Bug Fix:
+1. Fix clicking footer trigger loading when no more state
+2. footer indicator shouldn't hide when state in noMore,failed and not full in one page
+
+other:
+1. Remove asSliver usage in all indicators(no need to use,not support first sliver)
+2. make indicator auto fit boxSize,just like SliverToBoxAdapter